@@ -13,8 +13,8 @@ import 'util.dart';
 
 void main() {
   group('test build_example_command', () {
-    CommandRunner<void> runner;
-    RecordingProcessRunner processRunner;
+    late CommandRunner<void> runner;
+    late RecordingProcessRunner processRunner;
     final String flutterCommand =
         const LocalPlatform().isWindows ? 'flutter.bat' : 'flutter';
 
