@@ -44,7 +44,7 @@ class TestCommand extends PluginCommand {
 
       print('RUNNING $packageName tests...');
 
-      final String enableExperiment = argResults[kEnableExperiment] as String;
+      final String enableExperiment = getStringArg(kEnableExperiment);
 
       // `flutter test` automatically gets packages.  `pub run test` does not.  :(
       int exitCode = 0;
